restructure pactCreator and update imports
