+++ packages/libs/client/src/index.ts
@@ -9,3 +9,4 @@ export * from './signing-api/v1/sign';
 export * from './interfaces/IPactCommand';
 export * from './buildCommandFromTemplate';
 export * from './buildUnsignedTransaction';
+export * from './pactCreator';
