+++ packages/libs/client/src/tests/pact.test.ts
@@ -8,7 +8,8 @@ import { convertIUnsignedTransactionToNoSig } from '@kadena/chainweb-node-client
 import { PactNumber } from '@kadena/pactjs';
 import { IUnsignedCommand } from '@kadena/types';
 
-import { Pact, PactCommand } from '../pact';
+import { PactCommand } from '../pact';
+import { Pact } from '../pactCreator';
 
 import fetch from 'cross-fetch';
 const testURL: string = 'http://fake-api-host.local.co';
